Minor fixes for "Fairy Floss" theme in 1.1.1 (#1351)
My instance was updated to 1.1.1 recently, and I noticed the background
color does not fill the whole card for links without big preview images
on top when using the fairy-floss theme:


![image](https://github.com/hometown-fork/hometown/assets/525780/c184f3d7-58cc-4593-937b-7a167e53a0be)

Similarly, there's some background code missing to the right when the
link content (title / preview) don't stretch the card content to fill
the link horizontally:


![image](https://github.com/hometown-fork/hometown/assets/525780/d908f94f-1dce-446b-b795-25be4421c0fa)

This diff should fix that. There's some remaining styling for
`.status-card.compact`, but I don't know where that's used.

The second commit fixes some issues with the dropdown that appears when
clicking the "…" button on the bottom left of a post:

**Before**

![image](https://github.com/hometown-fork/hometown/assets/525780/5ef58f6a-f90a-4320-b122-55418b51227d)

**After**

![image](https://github.com/hometown-fork/hometown/assets/525780/fb4f89cf-817f-478b-a02c-a6d15e66bb49)

My apologies if these changes should be made upstream somewhere, but I
didn't find another repo for the theme.

---------

Co-authored-by: arne <noreply@null.arnes.space>