Fixing federation bug
There was a major issue where remote posts weren't being signed
correctly. Or rather, the signing process wanted access to a constant
that was defined in the wrong file. This is probably due to a
backporting error since this is precisely the files that were part of
the big security patches.