Getting rid of string/boolean confusion
The issue is in the dropdown_selector component, which stores data as a
`data-index` element on a dropdown. This means all values are cast to
strings in anything using that dropdown, which means we need to
unstringify on the `handleChange` function that (up in the component)
calls the `getAttribue` function. Anyway this means we can remove
terrible hacks in multiple places.